Revert "Rename navbar examples file"
This reverts commit 588d3391a2cb713dd79e3650bddaa39ea3381e94.
