+++ packages/docs-app/src/examples/core-examples/navbarExample.tsx
